fix: move to message_build.go
