+++ session.go
@@ -98,16 +98,6 @@ func (s *session) waitForInSessionTime() {
 	}
 }
 
-func insertSendingTime(msg *Message, sessionID SessionID, timestampPrecision TimestampPrecision) {
-	sendingTime := time.Now().UTC()
-
-	if sessionID.BeginString >= BeginStringFIX42 {
-		msg.Header.SetField(tagSendingTime, FIXUTCTimestamp{Time: sendingTime, Precision: timestampPrecision})
-	} else {
-		msg.Header.SetField(tagSendingTime, FIXUTCTimestamp{Time: sendingTime, Precision: Seconds})
-	}
-}
-
 func (s *session) insertSendingTime(msg *Message) {
 	insertSendingTime(msg, s.sessionID, s.timestampPrecision)
 }
@@ -116,32 +106,6 @@ func optionallySetID(msg *Message, field Tag, value string) {
 	if len(value) != 0 {
 		msg.Header.SetString(field, value)
 	}
-}
-
-func fillDefaultHeader(store MessageStore, msg *Message, inReplyTo *Message, sessionID SessionID, enableLastMsgSeqNumProcessed bool, timestampPrecision TimestampPrecision) error {
-	msg.Header.SetString(tagBeginString, sessionID.BeginString)
-	msg.Header.SetString(tagSenderCompID, sessionID.SenderCompID)
-	optionallySetID(msg, tagSenderSubID, sessionID.SenderSubID)
-	optionallySetID(msg, tagSenderLocationID, sessionID.SenderLocationID)
-
-	msg.Header.SetString(tagTargetCompID, sessionID.TargetCompID)
-	optionallySetID(msg, tagTargetSubID, sessionID.TargetSubID)
-	optionallySetID(msg, tagTargetLocationID, sessionID.TargetLocationID)
-
-	insertSendingTime(msg, sessionID, timestampPrecision)
-
-	if enableLastMsgSeqNumProcessed {
-		if inReplyTo != nil {
-			lastSeqNum, err := inReplyTo.Header.GetInt(tagMsgSeqNum)
-			if err != nil {
-				return err
-			}
-			msg.Header.SetInt(tagLastMsgSeqNumProcessed, lastSeqNum)
-		} else {
-			msg.Header.SetInt(tagLastMsgSeqNumProcessed, store.NextTargetMsgSeqNum()-1)
-		}
-	}
-	return nil
 }
 
 func (s *session) fillDefaultHeader(msg *Message, inReplyTo *Message) {
@@ -299,7 +263,7 @@ func (s *session) dropAndSendInReplyTo(msg *Message, inReplyTo *Message) error {
 }
 
 func (s *session) prepMessageForSend(msg *Message, inReplyTo *Message) (msgBytes []byte, err error) {
-	data := MessageBuildData{
+	data := BuildMessageInput{
 		Msg:                          msg,
 		InReplyTo:                    inReplyTo,
 		SessionID:                    s.sessionID,
@@ -308,7 +272,7 @@ func (s *session) prepMessageForSend(msg *Message, inReplyTo *Message) (msgBytes
 		EnableLastMsgSeqNumProcessed: s.SessionSettings.EnableLastMsgSeqNumProcessed,
 		TimestampPrecision:           s.timestampPrecision,
 	}
-	var output *MessageBuildOutputData
+	var output *BuildMessageOutput
 
 	if !s.DisableMessagePersist {
 		output, err = s.store.SaveMessageWithTx(&data)
@@ -327,56 +291,6 @@ func (s *session) prepMessageForSend(msg *Message, inReplyTo *Message) (msgBytes
 		return
 	}
 	msgBytes = output.MsgBytes
-	return
-}
-
-func buildMessage(store MessageStore, bd *MessageBuildData) (output *MessageBuildOutputData, err error) {
-	msg := bd.Msg
-	tmpErr := fillDefaultHeader(store, msg, bd.InReplyTo, bd.SessionID, bd.EnableLastMsgSeqNumProcessed, bd.TimestampPrecision)
-	if tmpErr != nil && bd.logger != nil {
-		(*bd.logger).OnEvent(err.Error())
-	}
-	outputData := MessageBuildOutputData{}
-	outputData.SeqNum = store.NextSenderMsgSeqNum()
-	msg.Header.SetField(tagMsgSeqNum, FIXInt(outputData.SeqNum))
-
-	msgType, err := msg.Header.GetBytes(tagMsgType)
-	if err != nil {
-		return
-	}
-
-	if isAdminMessageType(msgType) {
-		if bd.application != nil {
-			(*bd.application).ToAdmin(msg, bd.SessionID)
-		}
-
-		if bytes.Equal(msgType, msgTypeLogon) {
-			var resetSeqNumFlag FIXBoolean
-			if msg.Body.Has(tagResetSeqNumFlag) {
-				if err = msg.Body.GetField(tagResetSeqNumFlag, &resetSeqNumFlag); err != nil {
-					return
-				}
-			}
-
-			if resetSeqNumFlag.Bool() {
-				if err = store.Reset(); err != nil {
-					return
-				}
-
-				outputData.SentReset = true
-				outputData.SeqNum = store.NextSenderMsgSeqNum()
-				msg.Header.SetField(tagMsgSeqNum, FIXInt(outputData.SeqNum))
-			}
-		}
-	} else if bd.application != nil {
-		if err = (*bd.application).ToApp(msg, bd.SessionID); err != nil {
-			return
-		}
-	}
-
-	outputData.MsgBytes = msg.build()
-	output = &outputData
-
 	return
 }
